+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Chen 2021 lorlatinib exposure-response models ([doi:10.1002/cpt.2228](https://doi.org/10.1002/cpt.2228)) -- adults with ALK-positive or ROS1-positive non-small cell lung cancer.
+- Add Nakashima 2015 valproic acid seizure-control exposure-response models ([doi:10.1371/journal.pone.0141266](https://doi.org/10.1371/journal.pone.0141266)) -- Japanese children and adults with epilepsy.
 
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.